Extensible `SimF`
* Let `SimF` decide about `probe` (`Probe` type) `time` (`TimeDuraction`
  type) and `mvar` (`MVar` type) types.  For that the type signature of
 `SimF` contains `probe`, `time`, `mvar` variables.  This way `SimF` can
  decide how the simulation environment looks like.
* Provide `MonadSay`, `MonadProbe`, `MonadConc` instance for
  `Free (SimF ...)` and `Free (Sum (SimF ...) g)`
  which are generic in the `probe`, `time` and `mvar` variables.
* Provide instances for `MonadSay IO`, `MonadProbe IO` and `MonadConc IO`

This allows to build expressions in `(Monad m, MonadSay m, MonadProbe m,
MonadConc m) => m a` and they can be interpreted in
* `Free (SimF ...)`
* `Free (Sum (SimF ...) (SendRecvF ...))`
* `IO`

The benefit of this approach is that we don't have any runtime overhead
when we will will run `(MonadSay m, MonadProbe m, MonadConc m) => m a`
in `IO`.

Finally `SimF` is extensible: i.e. it can be extended to `Sum (SimF ...)
(SendRecvF ...)` and the `SendRecvF ...` effects can be run inside
`Fork` or `Timer` directives.

The price we pay is `SimF` type signature:
```
data SimF (g :: * -> *) (probe :: * -> * -> *) time (mvar :: * -> *) s a where
```